+++ stroom-security/stroom-security-impl/src/main/java/stroom/security/impl/session/SessionListListener.java
@@ -111,7 +111,7 @@ class SessionListListener implements HttpSessionListener, SessionListService {
         } else {
             // This is a different node so make a rest call to it to get the result
 
-            final String url = NodeCallUtil.getBaseEndpointUrl(nodeService, nodeName) +
+            final String url = NodeCallUtil.getBaseEndpointUrl(nodeInfo, nodeService, nodeName) +
                     ResourcePaths.buildAuthenticatedApiPath(
                             SessionResource.BASE_PATH,
                             SessionResource.LIST_PATH_PART);